requirements.txt
agate==1.6.3
asn1crypto==1.5.1
attrs==22.1.0
Babel==2.10.3
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.0.0
colorama==0.4.5
cryptography==36.0.2
dbt-core==1.3.1
dbt-extractor==0.4.1
dbt-snowflake==1.3.0
filelock==3.8.0
future==0.18.2
hologram==0.0.15
idna==3.3
importlib-metadata==4.12.0
isodate==0.6.1
jaraco.classes==3.2.2
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==3.2.0
keyring==23.9.1
leather==0.3.4
Logbook==1.5.3
MarkupSafe==2.1.1
mashumaro==3.0.4
minimal-snowplow-tracker==0.0.2
more-itertools==8.14.0
msgpack==1.0.4
networkx==2.8.6
oscrypto==1.3.0
packaging==21.3
parsedatetime==2.4
pycparser==2.21
pycryptodomex==3.15.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytimeparse==1.1.8
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
SecretStorage==3.3.3
six==1.16.0
snowflake-connector-python==2.7.12
snowflake-snowpark-python==0.9.0
sqlparse==0.4.2
text-unidecode==1.3
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
